Optimises delays of "Complete" HUDs

+++ piwigo/Loading View/LoadingView.m
@@ -124,7 +124,7 @@
 		self.progressLabel.alpha = 0;
 	}];
 	
-	dispatch_after(dispatch_time(DISPATCH_TIME_NOW, delay * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
+	dispatch_after(dispatch_time(DISPATCH_TIME_NOW, (int64_t)(delay * NSEC_PER_SEC)), dispatch_get_main_queue(), ^{
 		[UIView animateWithDuration:0.3 animations:^{
 			weakSelf.alpha = 0;
 		} completion:^(BOOL finished) {
